+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/__mocks__/useSelectAssetFlow.mock.ts
@@ -1,68 +1,14 @@
 import { AssetType } from "@ledgerhq/react-ui/pre-ldls/index";
-import { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import type { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/wallet-api/ModularDrawer/types";
+import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets/index";
+import { TokenCurrency } from "@ledgerhq/types-cryptoassets";
 
-export const bitcoinCurrency: CryptoCurrency = {
-  type: "CryptoCurrency",
-  id: "bitcoin",
-  coinType: 60,
-  name: "Bitcoin",
-  managerAppName: "Bitcoin",
-  ticker: "BTC",
-  scheme: "bitcoin",
-  color: "#ffae35",
-  family: "bitcoin",
-  units: [
-    {
-      name: "bitcoin",
-      code: "BTC",
-      magnitude: 8,
-    },
-  ],
-  explorerId: "btc",
-  explorerViews: [],
-};
-
-export const ethereumCurrency: CryptoCurrency = {
-  type: "CryptoCurrency",
-  id: "ethereum",
-  coinType: 60,
-  name: "Ethereum",
-  managerAppName: "Ethereum",
-  ticker: "ETH",
-  scheme: "ethereum",
-  color: "#0ebdcd",
-  family: "evm",
-  units: [
-    {
-      name: "ether",
-      code: "ETH",
-      magnitude: 18,
-    },
-  ],
-  explorerId: "eth",
-  explorerViews: [],
-};
-
-export const arbitrumCurrency = {
-  type: "CryptoCurrency" as const,
-  id: "arbitrum" as const,
-  coinType: 60,
-  name: "Arbitrum",
-  managerAppName: "Ethereum",
-  ticker: "ETH",
-  scheme: "arbitrum",
-  color: "#28a0f0",
-  family: "evm",
-  units: [
-    {
-      name: "ETH",
-      code: "ETH",
-      magnitude: 18,
-    },
-  ],
-  explorerViews: [],
-};
+export const bitcoinCurrency = getCryptoCurrencyById("bitcoin");
+export const ethereumCurrency = getCryptoCurrencyById("ethereum");
+export const arbitrumCurrency = getCryptoCurrencyById("arbitrum");
+export const baseCurrency = getCryptoCurrencyById("base");
+export const scrollCurrency = getCryptoCurrencyById("scroll");
+export const injectiveCurrency = getCryptoCurrencyById("injective");
 
 export const arbitrumToken: TokenCurrency = {
   type: "TokenCurrency",
@@ -77,6 +23,22 @@ export const arbitrumToken: TokenCurrency = {
       name: "Arbitrum",
       code: "ARB",
       magnitude: 18,
+    },
+  ],
+};
+export const usdcToken: TokenCurrency = {
+  type: "TokenCurrency",
+  id: "ethereum/erc20/usd__coin",
+  contractAddress: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+  parentCurrency: ethereumCurrency,
+  tokenType: "erc20",
+  name: "USD Coin",
+  ticker: "USDC",
+  units: [
+    {
+      name: "USD Coin",
+      code: "USDC",
+      magnitude: 6,
     },
   ],
 };
@@ -101,6 +63,8 @@ export const ethereumAssetType: AssetType = {
 
 export const findCryptoCurrencyById = (id: string) =>
   [bitcoinCurrency, ethereumCurrency, arbitrumCurrency].find(a => a.id === id);
+export const getTokenOrCryptoCurrencyById = (id: string) =>
+  [bitcoinCurrency, ethereumCurrency, arbitrumCurrency, arbitrumToken].find(a => a.id === id);
 
 export const mockAssetsConfiguration: EnhancedModularDrawerConfiguration["assets"] = {
   filter: "topNetworks",