+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/Form/SwapWebViewDemo3.tsx
@@ -55,6 +55,7 @@ import {
 } from "../utils/index";
 import FeesDrawerLiveApp from "./FeesDrawerLiveApp";
 import WebviewErrorDrawer from "./WebviewErrorDrawer/index";
+import { currentRouteNameRef } from "~/renderer/analytics/screenRefs";
 
 export class UnableToLoadSwapLiveError extends Error {
   constructor(message: string) {
@@ -510,6 +511,7 @@ const SwapWebView = ({ manifest }: SwapWebProps) => {
         <Web3AppWebview
           manifest={manifestWithHash}
           inputs={{
+            source: currentRouteNameRef.current || "",
             theme: themeType,
             lang: locale,
             currencyTicker: fiatCurrency.ticker,