+++ apps/ledger-live-mobile/src/navigation/DeeplinksProvider.tsx
@@ -34,6 +34,7 @@ import { navigationIntegration } from "../sentry";
 import { OptionMetadata } from "~/reducers/types";
 const TRACKING_EVENT = "deeplink_clicked";
 import { DdRumReactNavigationTracking } from "@datadog/mobile-react-navigation";
+import { viewNamePredicate } from "~/datadog";
 
 const themes: {
   [key: string]: Theme;
@@ -700,7 +701,7 @@ export const DeeplinksProvider = ({
         (isReadyRef as Writeable<typeof isReadyRef>).current = true;
         setTimeout(() => SplashScreen.hide(), 300);
         navigationIntegration.registerNavigationContainer(navigationRef);
-        DdRumReactNavigationTracking.startTrackingViews(navigationRef.current);
+        DdRumReactNavigationTracking.startTrackingViews(navigationRef.current, viewNamePredicate);
       }}
     >
       {children}