+++ e2e/mobile/specs/swap/otherTestCases/swap_noAccountTo.spec.ts
@@ -5,6 +5,7 @@ const noAccountToTestConfig = {
   account2: Account.BTC_NATIVE_SEGWIT_1,
   testTitle: "from Account present to Account not present",
   tmsLinks: ["B2CQA-3354"],
+  tags: ["@NanoSP", "@LNS", "@NanoX"],
 };
 
 runSwapWithoutAccountTest(
@@ -13,5 +14,5 @@ runSwapWithoutAccountTest(
   noAccountToTestConfig.testTitle,
   noAccountToTestConfig.tmsLinks,
   "noAccountTo",
-  ["@NanoSP", "@LNS", "@NanoX"],
+  noAccountToTestConfig.tags,
 );