+++ libs/ledgerjs/packages/swift-bridge-hw-transport-ble/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/swift-bridge-hw-transport-ble
 
+## 1.2.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.7
+
+## 1.2.7-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.7-next.0
+
 ## 1.2.6
 
 ### Patch Changes